requirements.txt
asgiref==3.11.1
Django==6.0.4
openpyxl==3.1.5
python-dotenv==1.2.2
sqlparse==0.5.5
